Remove slave configurator and slave client
